feat: 마이 인포 / 마이아티스트인포 / 아티스트인포 dto 수정 완료

+++ src/main/java/com/server/objet/global/repository/FollowingRepository.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 
 public interface FollowingRepository extends JpaRepository<Follow, Long> {
-    Long countByUserId(Long userId);
-    Long countByArtistId(Long artistId);
+    int countByUserId(Long userId);
+    int countByArtistId(Long artistId);
 
     Optional<Follow> findByUserIdAndArtistId(Long userId, Long artistId);
 }
